+++ CHANGELOG.md
@@ -26,6 +26,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Previous acme.sh certs migration
 - "wo maintenance" command to perform server package update & cleanup
 - Support for Netdata on backend : https://server.hostname:22222/netdata/
+- Stack : composer and netdata
 
 #### Changed
 
@@ -51,6 +52,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - EasyEngine database sync during migration
 - command "wo info"
 - phpmyadmin install with composer
+- command "wo clean --memcached"
 
 ### v3.9.4 - 2019-03-15